requirements.txt
asgiref==3.10.0
Django==5.2.7
django-crispy-forms==2.4
django-filter==25.2
djangorestframework==3.16.1
pillow==11.3.0
psycopg2-binary==2.9.11
sqlparse==0.5.3
typing_extensions==4.15.0
